Add dm prune
